added option to not show the alert dialog that the icon changed
This is the only way I found to not show the alert in Objective C. I added the option from flutter so the developer can choose wether or not s/he wants to show it.

I couldn't find a way to send a bool through the invoke method, that is why i did it as a string. Maybe you know a better way.

For all I've seen this won't get your app rejected by the app store review. But I left the option in case someone faces the issue and can easily change it back to the basic configuration.